# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Quorum SQL API
  version: 1.0.0
  description: REST API for Quorum SQL server.
servers:
  - url: http://localhost:8787
paths:
...
  /queries:
    get:
      summary: List queries
      parameters:
        - in: query
          name: teamId
          required: true
          schema: { type: string }
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200': { description: OK }
    post:
      summary: Create query
      responses:
        '200': { description: OK }
  /queries/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Get query
      responses:
        '200': { description: OK }
    patch:
      summary: Update query
      responses:
        '200': { description: OK }
    delete:
      summary: Delete query
      responses:
        '200': { description: OK }
...